Made dependency versions specific for future stability
The current repository doesn't compile on my machine, propably because the dependencies developed further without backwards compatibility. I searched for the newest dependecy versions at the time when you created the repository and made them explicit. Now it compiles on my machine and should be stable even in the future.